+++ lib/bali/form_builder/email_fields.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Bali
+  class FormBuilder < ActionView::Helpers::FormBuilder
+    module EmailFields
+      def email_field_group(method, options = {})
+        @template.render(Bali::FieldGroupWrapper::Component.new(self, method, options)) do
+          email_field(method, options)
+        end
+      end
+
+      def email_field(method, options = {})
+        field_helper(method, super(method, field_options(method, options)), options)
+      end
+    end
+  end
+end