+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.15.2] - 2022-06-30
+
+- Improve `FormBuilder` testing.
+
+## [0.15.1] - 2022-06-29
+
+- Pass the `options` parameter to the `SideMenu::Item::Component`.
+
+## [0.15.0] - 2022-06-28
+
+- Added `FormBuilder` and `FieldGroupWrapperComponent`.
+
 ## [0.14.0] - 2022-06-28
 
 - Added `SideMenu::Component`.